+++ src/module/data/helpers.mjs
@@ -7,9 +7,10 @@ const { NumberField, SchemaField, StringField } = foundry.data.fields;
  * @param {number} [options.min=0]      The minimum value for the field.
  * @param {number} [options.max]        The maximum value for the field.
  * @param {string} [options.label]      Label for the field.
+ * @param {boolean} [options.persisted] Is the field saved to the database?
  * @returns A number field that is non-nullable and always defined.
  */
-export const requiredInteger = ({ initial = 0, min = 0, max, label } = {}) => new NumberField({ initial, label, min, max, required: true, nullable: false, integer: true });
+export const requiredInteger = ({ initial = 0, min = 0, max, label, persisted } = {}) => new NumberField({ initial, label, min, max, persisted, required: true, nullable: false, integer: true });
 
 /* -------------------------------------------------- */