+++ src/module/data/item/ability.mjs
@@ -1,5 +1,5 @@
 import { DrawSteelActiveEffect, DrawSteelChatMessage } from "../../documents/_module.mjs";
-import { setOptions, validateDSID } from "../helpers.mjs";
+import { requiredInteger, setOptions, validateDSID } from "../helpers.mjs";
 import BaseItemModel from "./base-item.mjs";
 import DamagePowerRollEffect from "../pseudo-documents/power-roll-effects/damage-effect.mjs";
 import FormulaField from "../fields/formula-field.mjs";
@@ -87,6 +87,11 @@ export default class AbilityModel extends BaseItemModel {
         characteristics: new fields.SetField(setOptions()),
       }),
       effects: new ds.data.fields.CollectionField(ds.data.pseudoDocuments.powerRollEffects.BasePowerRollEffect),
+
+      characteristic: new fields.SchemaField({
+        key: new fields.StringField({ choices: Object.keys(ds.CONFIG.characteristics) }),
+        value: requiredInteger({ initial: -5, min: null }),
+      }, { persisted: false }),
     });
 
     schema.effect = new fields.SchemaField({
@@ -103,33 +108,12 @@ export default class AbilityModel extends BaseItemModel {
 
   /* -------------------------------------------------- */
 
-  /**
-   * Helper for _applyAbilityBonuses.
-   * TODO: Replace with non-persisted fields in v14.
-   */
-  static #forcedBonus = new fields.NumberField();
-
-  /* -------------------------------------------------- */
-
   /** @inheritdoc */
   static migrateData(data) {
     // Game release updates
     if (data.type === "action") data.type = "main";
 
     return super.migrateData(data);
-  }
-
-  /* -------------------------------------------------- */
-
-  /** @inheritdoc */
-  prepareBaseData() {
-    super.prepareBaseData();
-    for (const effect of this.power.effects) effect.prepareBaseData();
-
-    this.power.characteristic = {
-      key: "",
-      value: -5,
-    };
   }
 
   /* -------------------------------------------- */
@@ -244,8 +228,9 @@ export default class AbilityModel extends BaseItemModel {
         // Apply forced movement bonuses to all forced movement effects
         for (const effect of this.power.effects) {
           if (effect.type !== "forced") continue;
+          const numberField = effect.schema.getField(key);
           const currentBonus = foundry.utils.getProperty(effect, key) ?? 0;
-          foundry.utils.setProperty(effect, key, AbilityModel.#forcedBonus.applyChange(currentBonus, this, bonus));
+          foundry.utils.setProperty(effect, key, numberField.applyChange(currentBonus, this, bonus));
         }
       }